refactor(signup form): create SignUpForm and update SignUpPage
