+++ src/pages/StartPage/index.tsx
@@ -8,7 +8,6 @@ import TwitterLogo from '@/assets/image/icons/twitter-logo.svg';
 import { Footer } from '@/components/Footer';
 import { Button } from '@/components/ui-components/Button';
 import { ButtonStyled1 } from '@/components/ui-components/Button/config';
-import { ContentText } from '@/components/ui-components/ContentText';
 import { Icon } from '@/components/ui-components/Icon';
 import { StyledIcon40 } from '@/components/ui-components/Icon/config';
 import { LinkApp } from '@/components/ui-components/Link';
@@ -18,9 +17,10 @@ import { PATH } from '@/constants/routerLinks';
 import { createAccountWithGoogle } from '@/services/auth/createUserWithGoogle';
 import { getUserDataFromFirestore } from '@/services/firestore/getUserDataFromFirestore';
 import { setUser } from '@/store/sliceUser';
-import { FONT_SIZE } from '@/theme/variables';
 
-import { ContainerWrapper, Content, Image } from './styles';
+import {
+  ContainerWrapper, Content, ContentLinks, ContentText, Image,
+} from './styles';
 
 const StartPage = () => {
   const {
@@ -74,10 +74,7 @@ const StartPage = () => {
           >
             {SIGN_UP_EMAIL}
           </Button>
-          <ContentText
-            size={FONT_SIZE.xs}
-            width="375px"
-          >
+          <ContentText>
             {'By singing up you agree to the '}
             <LinkApp>{LINKS_TO.service}</LinkApp>
             {' and '}
@@ -85,11 +82,11 @@ const StartPage = () => {
             {', including '}
             <LinkApp>{LINKS_TO.cookie}</LinkApp>
           </ContentText>
-          <ContentText size={FONT_SIZE.sm}>
+          <ContentLinks>
             <span>{ACCOUNT}</span>
             {' '}
             <LinkApp to={PATH.LOG_IN_PAGE}>{LINKS_TO.login}</LinkApp>
-          </ContentText>
+          </ContentLinks>
         </Content>
       </ContainerWrapper>
       <Footer />
