+++ src/pages/StartPage/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { BREAKPOINTS, SPACING } from '@/theme/variables';
+import { BREAKPOINTS, FONT_SIZE, SPACING } from '@/theme/variables';
 
 export const ContainerWrapper = styled.div`
   display: flex;
@@ -33,4 +33,15 @@ export const Content = styled.div`
       width: 100%;
     }
   }
+`;
+
+export const ContentText = styled.div`
+  width: 375px;
+  font-size: ${FONT_SIZE.xs};
+  margin-bottom: ${SPACING.sm};
+`;
+
+export const ContentLinks = styled.div`
+  font-size: ${FONT_SIZE.sm};
+  margin-bottom: ${SPACING.sm};
 `;
